requirements.txt
# ── Core Deep Learning ─────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── Data Science ───────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ── Computer Vision ────────────────────────────────────────────
Pillow>=10.0.0
opencv-python>=4.8.0

# ── Visualization ──────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── API Backend (for React UI) ─────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0

# ── Utilities ─────────────────────────────────────────────────
tqdm>=4.65.0
